pyproject.toml
[tool.poetry]
name = "ocr-model-repo-script"
version = "0.1.0"
description = "Scripts to build gh-pages for ocr-model-repos and catalogues"
authors = ["JKamlah <jan.kamlah@uni-mannheim.de>"]
license = "Attribution-ShareAlike 4.0 International"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
rich = "^13.7.0"
PyYAML = "^6.0.1"
lxml = "^5.1.0"
GitPython = "^3.1.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
